Extract CShapes data and matrix

Functions to import CShapes 2.0 datasets and distances from the [cShapes] package and format them to a qVerse consistent output for creating maps.

import_cshapes(date, ...)

import_distlist(date, type, ...)

import_distmatrix(date, type, ...)

Arguments

date
The date for which the distance list should be computed. This argument must be of type Date and must be in the range 1/1/1886 - end of the dataset.

...
Arguments to be passed to [cshapes] functions. See the [cshapes] documentation for more details

type
Specifies the type of distance list: "capdist" for capital distances, "centdist" for centroid distances, and "mindist" for minimum distances.

Value

A dataframe with the qVerse-consistently formatted [cshapes] dataset.

A dataframe with the desired distance list between polygons, capitals, or polygon centroids in kilometers.

A matrix with the desired distance list between polygons, capitals, or polygon centroids in kilometers.

Details

import_cshapes()imports CShapes 2.0 datasets and formats them to a qVerse consistent output.

import_distlist() imports pre-computed minimum distance dataframe from the [cShapes] package. Minimum distances are computed in three ways: distances between capitals, distances between centroids of the polygons, minimum distances between the polygons in kilometers.

import_distlist() imports and formats pre-computed minimum distance matrix from the [cShapes] package. Minimum distances are computed in three ways: distances between capitals, distances between centroids of the polygons, minimum distances between the polygons in kilometers.

Examples

# \donttest{
import_cshapes(date = "1900-01-01")

#> # A tibble: 42 x 11
#>    COW_Nr Beg        End        Label                  geometry Capital CapitalLong
#>    <chr>  <messydt>  <messydt>  <chr>        <MULTIPOLYGON [°]> <chr>         <dbl>
#>  1 100    1886-01-01 1903-11-02 Colo~ (((-78.1309 2.500494, -7~ Bogota      -74.1  
#>  2 101    1886-01-01 2019-12-31 Vene~ (((-60.82334 8.652364, -~ Caracas     -66.9  
#>  3 130    1886-01-01 1942-01-28 Ecua~ (((-78.80972 1.437782, -~ Quito       -78.5  
#>  4 135    1886-01-01 1909-07-08 Peru  (((-69.45264 -1.526036, ~ Lima        -77.1  
#>  5 140    1886-01-01 1903-11-16 Braz~ (((-53.3984 -33.74904, -~ Rio De~     -43.2  
#>  6 145    1886-01-01 1903-11-16 Boli~ (((-58.15889 -20.16805, ~ La Paz      -68.2  
#>  7 150    1886-01-01 1938-07-20 Para~ (((-54.59892 -25.57323, ~ Asunci~     -57.7  
#>  8 165    1886-01-01 2019-12-31 Urug~ (((-53.3984 -33.74904, -~ Montev~     -56.2  
#>  9 2      1886-01-01 1959-01-02 Unit~ (((-75.48306 35.34583, -~ Washin~     -77.0  
#> 10 200    1886-01-01 1921-12-05 Unit~ (((-7.252509 55.07059, -~ London       -0.117
#> # ... with 32 more rows, and 4 more variables: CapitalLat <dbl>,
#> #   WellDefinedBorders <int>, Status <dbl>, Owner <chr>

# }
# \donttest{
import_distlist(date = "1900-01-01", type = "capdist")

#> # A tibble: 1,764 x 5
#>    FromLabel     FromCode ToLabel            ToCode Distance
#>    <chr>            <dbl> <chr>               <dbl>    <dbl>
#>  1 United States        2 United States           2       0 
#>  2 United States        2 Haiti                  41    2302.
#>  3 United States        2 Dominican Republic     42    2367.
#>  4 United States        2 Mexico                 70    3029.
#>  5 United States        2 Guatemala              90    2998.
#>  6 United States        2 Honduras               91    2924.
#>  7 United States        2 El Salvador            92    3037.
#>  8 United States        2 Nicaragua              93    3101.
#>  9 United States        2 Colombia              100    3810.
#> 10 United States        2 Venezuela             101    3303.
#> # ... with 1,754 more rows

# }
# \donttest{
import_distmatrix(date = "1900-01-01", type = "capdist")

#>             2         41         42        70         90         91         92
#> 2       0.000  2301.5532  2367.1128  3029.181  2997.6396  2923.9591  3036.7186
#> 41   2301.553     0.0000   257.2608  2821.354  1988.1533  1663.6216  1880.7351
#> 42   2367.113   257.2608     0.0000  3077.711  2241.0005  1911.7989  2129.7961
#> 70   3029.181  2821.3543  3077.7106     0.000  1060.0142  1400.7600  1234.6813
#> 90   2997.640  1988.1533  2241.0005  1060.014     0.0000   361.6788   174.9823
#> 91   2923.959  1663.6216  1911.7989  1400.760   361.6788     0.0000   219.0084
#> 92   3036.719  1880.7351  2129.7961  1234.681   174.9823   219.0084     0.0000
#> 93   3100.647  1653.6023  1890.2908  1596.043   536.1107   238.8973   362.0891
#> 100  3809.909  1553.6794  1600.0512  3175.648  2115.8511  1783.2672  1941.6265
#> 101  3302.757  1063.7654   938.2014  3597.095  2604.4918  2242.9431  2450.9058
#> 130  4333.377  2181.4369  2270.6308  3130.631  2108.7805  1851.9763  1940.2638
#> 135  5639.440  3422.6283  3465.3371  4241.888  3302.9350  3101.9104  3148.6673
#> 140  7697.051  5573.5604  5420.6976  7673.364  6620.0487  6306.2558  6445.3124
#> 145  6201.384  3902.6250  3872.3234  5224.664  4230.5785  3981.1413  4065.0036
#> 150  7383.562  5100.1342  5016.4650  6679.080  5672.7020  5407.8697  5504.3896
#> 155  8036.464  5756.0184  5745.8366  6587.653  5722.7144  5547.8667  5577.6115
#> 160  8353.007  6054.8735  5990.6218  7353.785  6404.0388  6176.4295  6244.2751
#> 165  8443.935  6151.9843  6077.6134  7532.264  6570.0382  6333.8799  6408.1986
#> 200  5914.798  7189.6463  7026.8094  8943.132  8777.9335  8594.8896  8765.0448
#> 210  6208.484  7535.7074  7375.9463  9232.289  9098.3364  8922.9870  9089.8159
#> 211  6232.520  7501.0831  7335.5610  9260.297  9097.6557  8913.6602  9084.3329
#> 220  6179.954  7363.8406  7190.9803  9208.883  9004.7112  8808.8518  8984.6597
#> 225  6618.562  7760.6695  7581.4827  9647.104  9429.3776  9227.3880  9405.9273
#> 230  6106.121  6907.0644  6705.1383  9079.549  8708.4048  8466.3814  8660.7727
#> 235  5751.461  6438.8341  6231.7589  8685.986  8270.1798  8017.8740  8215.9588
#> 255  6728.711  8108.3975  7950.6767  9741.803  9645.5637  9478.7142  9641.8744
#> 300  7142.736  8400.4201  8228.4775 10167.859 10015.5795  9829.8627 10001.5631
#> 325  7234.684  8237.4605  8043.6041 10256.246  9979.8934  9758.4894  9945.1658
#> 345  7608.974  8794.6110  8613.6919 10636.957 10455.1194 10258.4504 10435.0117
#> 350  8273.466  9276.8453  9077.9212 11299.639 11033.8771 10810.5298 10998.2351
#> 360  7997.778  9234.0354  9055.6161 11020.642 10871.9086 10682.7841 10856.2185
#> 365  7214.968  8923.1284  8801.5214 10100.789 10210.9513 10106.4152 10239.9187
#> 380  6654.107  8274.9656  8144.9463  9599.033  9637.2830  9512.7024  9655.9121
#> 390  6528.784  8008.0989  7862.1455  9523.555  9476.8255  9326.2157  9481.8197
#> 530 11540.680 11879.5118 11633.8069 14440.423 13854.8988 13542.8777 13759.0678
#> 600  6159.396  6653.7999  6432.9818  9032.568  8545.0124  8272.6753  8477.6864
#> 630 10202.901 11586.5298 11411.4072 13166.196 13154.8808 12993.2506 13155.2275
#> 640  8418.075  9586.8792  9400.4647 11444.670 11266.5898 11066.8619 11244.9490
#> 710 11169.617 13469.0679 13507.0425 12475.278 13357.2858 13546.3403 13506.3279
#> 730 11189.007 13474.4021 13556.1140 12068.443 13029.4632 13266.1635 13191.3395
#> 740 10926.086 13121.3837 13250.0327 11319.145 12336.5324 12614.5727 12506.9621
#> 800 14170.901 16348.9574 16291.2541 15756.301 16642.8030 16810.4034 16787.2360
#>             93        100        101        130       135       140       145
#> 2    3100.6468  3809.9095  3302.7573  4333.3768  5639.440  7697.051  6201.384
#> 41   1653.6023  1553.6794  1063.7654  2181.4369  3422.628  5573.560  3902.625
#> 42   1890.2908  1600.0512   938.2014  2270.6308  3465.337  5420.698  3872.323
#> 70   1596.0432  3175.6484  3597.0945  3130.6310  4241.888  7673.364  5224.664
#> 90    536.1107  2115.8511  2604.4918  2108.7805  3302.935  6620.049  4230.578
#> 91    238.8973  1783.2672  2242.9431  1851.9763  3101.910  6306.256  3981.141
#> 92    362.0891  1941.6265  2450.9058  1940.2638  3148.667  6445.312  4065.004
#> 93      0.0000  1579.7440  2119.9501  1614.7594  2863.615  6089.367  3744.030
#> 100  1579.7440     0.0000  1025.0413   724.4986  1870.302  4531.816  2423.276
#> 101  2119.9501  1025.0413     0.0000  1746.1146  2733.973  4510.433  2989.192
#> 130  1614.7594   724.4986  1746.1146     0.0000  1318.413  4570.673  2129.272
#> 135  2863.6150  1870.3019  2733.9726  1318.4131     0.000  3779.860  1078.893
#> 140  6089.3670  4531.8162  4510.4331  4570.6728  3779.860     0.000  2706.505
#> 145  3744.0298  2423.2760  2989.1921  2129.2724  1078.893  2706.505     0.000
#> 150  5173.4626  3750.9094  4082.0104  3564.1482  2508.662  1493.577  1458.826
#> 155  5311.0533  4226.6603  4880.0137  3768.8877  2457.479  2927.727  1894.594
#> 160  5937.8618  4629.0417  5065.3372  4328.6420  3112.539  1985.509  2215.753
#> 165  6095.8421  4753.8159  5146.3780  4482.7774  3290.918  1828.465  2358.124
#> 200  8695.5746  8501.8007  7500.1075  9223.9127 10161.949  9253.205  9986.451
#> 210  9028.4805  8857.5549  7857.3324  9579.1894 10520.802  9553.352 10340.050
#> 211  9013.2159  8800.0498  7794.2407  9523.1122 10441.293  9402.903 10235.998
#> 220  8900.1055  8632.1912  7620.4979  9356.3419 10246.088  9144.410 10014.637
#> 225  9313.5212  8995.6903  7978.0496  9720.1640 10566.855  9253.930 10281.481
#> 230  8521.5316  8024.3394  6999.4712  8744.6861  9503.874  8116.413  9160.392
#> 235  8064.5856  7529.2295  6504.1897  8248.3261  8999.700  7691.212  8662.576
#> 255  9589.1120  9434.4664  8433.6619 10156.1310 11091.617  9987.702 10888.824
#> 300  9926.9015  9664.3588  8649.6995 10388.7819 11250.169  9854.773 10961.929
#> 325  9828.8931  9387.7506  8363.1535 10108.8288 10857.426  9179.449 10468.357
#> 345 10347.0226 10011.2331  8990.0862 10735.1562 11535.777  9891.099 11178.299
#> 350 10877.8453 10382.7204  9358.0131 11098.7390 11768.131  9706.865 11284.497
#> 360 10776.1903 10459.4631  9438.6197 11183.4886 11981.522 10246.425 11609.310
#> 365 10253.3782 10360.6543  9403.1227 11064.4363 12123.986 11310.645 12050.004
#> 380  9649.1761  9691.8610  8725.3227 10400.4425 11440.246 10682.391 11357.370
#> 390  9447.2741  9377.1485  8390.3686 10094.1867 11080.069 10171.628 10937.403
#> 530 13517.8969 12435.4409 11526.7223 13018.3754 13018.386  9589.747 12093.935
#> 600  8299.8963  7635.2098  6614.7530  8343.2331  8994.445  7383.860  8567.310
#> 630 13101.8310 12806.6909 11783.1266 13529.2518 14242.603 11858.921 13721.893
#> 640 11152.0076 10765.8498  9741.5918 11487.5405 12223.954 10265.954 11786.982
#> 710 13784.4826 14959.4784 14410.8181 15356.5996 16648.013 17321.607 17371.001
#> 730 13504.1242 14855.5015 14490.7277 15117.5377 16310.221 18133.696 17243.568
#> 740 12845.6031 14327.5379 14180.9032 14443.8526 15493.886 18560.242 16534.338
#> 800 17045.3260 17891.3193 16991.1777 18503.4541 19700.806 16075.995 18777.986
#>           150       155        160        165        200        210        211
#> 2    7383.562  8036.464  8353.0069  8443.9351  5914.7982  6208.4845  6232.5202
#> 41   5100.134  5756.018  6054.8735  6151.9843  7189.6463  7535.7074  7501.0831
#> 42   5016.465  5745.837  5990.6218  6077.6134  7026.8094  7375.9463  7335.5610
#> 70   6679.080  6587.653  7353.7847  7532.2645  8943.1320  9232.2887  9260.2967
#> 90   5672.702  5722.714  6404.0388  6570.0382  8777.9335  9098.3364  9097.6557
#> 91   5407.870  5547.867  6176.4295  6333.8799  8594.8896  8922.9870  8913.6602
#> 92   5504.390  5577.611  6244.2751  6408.1986  8765.0448  9089.8159  9084.3329
#> 93   5173.463  5311.053  5937.8618  6095.8421  8695.5746  9028.4805  9013.2159
#> 100  3750.909  4226.660  4629.0417  4753.8159  8501.8007  8857.5549  8800.0498
#> 101  4082.010  4880.014  5065.3372  5146.3780  7500.1075  7857.3324  7794.2407
#> 130  3564.148  3768.888  4328.6420  4482.7774  9223.9127  9579.1894  9523.1122
#> 135  2508.662  2457.479  3112.5386  3290.9183 10161.9488 10520.8019 10441.2930
#> 140  1493.577  2927.727  1985.5093  1828.4645  9253.2048  9553.3524  9402.9034
#> 145  1458.826  1894.594  2215.7528  2358.1239  9986.4512 10340.0496 10235.9984
#> 150     0.000  1552.175  1037.7757  1072.9495 10188.8626 10519.5448 10386.7793
#> 155  1552.175     0.000  1114.3223  1344.6650 11650.6162 11991.5063 11867.7023
#> 160  1037.776  1114.322     0.0000   231.0916 11114.4948 11433.6057 11292.4034
#> 165  1072.949  1344.665   231.0916     0.0000 11017.6664 11331.1182 11186.6523
#> 200 10188.863 11650.616 11114.4948 11017.6664     0.0000   358.8551   319.9172
#> 210 10519.545 11991.506 11433.6057 11331.1182   358.8551     0.0000   173.5186
#> 211 10386.779 11867.702 11292.4034 11186.6523   319.9172   173.5186     0.0000
#> 220 10139.529 11627.320 11038.9992 10931.2348   341.1567   428.4049   261.7774
#> 225 10319.772 11832.536 11183.8397 11062.6862   751.0537   631.6566   492.3778
#> 230  9167.090 10681.863 10035.3238  9917.7985  1263.9856  1481.5680  1315.6183
#> 235  8701.100 10205.693  9587.9534  9478.0104  1585.2040  1863.8269  1711.6983
#> 255 11013.445 12506.219 11901.2602 11787.3126   933.5904   576.9399   654.0341
#> 300 10965.391 12489.093 11805.1182 11674.6813  1237.8726   936.1867   918.0077
#> 325 10351.692 11893.370 11148.3805 11006.5037  1433.8758  1294.2112  1172.8665
#> 345 11074.711 12616.852 11864.4391 11719.2298  1694.2508  1415.6560  1376.6605
#> 350 10995.182 12547.098 11692.1677 11525.1817  2394.3745  2163.6266  2092.0396
#> 360 11464.958 13012.452 12227.3922 12074.3299  2095.2433  1789.2937  1775.3262
#> 365 12293.192 13752.719 13207.0506 13100.9056  2105.3779  1775.0048  1914.7662
#> 380 11622.013 13068.377 12551.8806 12454.2882  1437.3864  1129.8899  1286.1218
#> 390 11141.328 12608.299 12056.3100 11953.0344   957.9557   622.8365   768.8093
#> 530 11082.688 12363.843 11256.4239 11029.7822  5892.0865  5698.0714  5609.1904
#> 600  8479.874 10010.768  9320.6401  9195.3570  2015.9849  2242.5219  2075.9597
#> 630 13261.920 14782.969 13795.1494 13596.0807  4408.8204  4073.3701  4092.9306
#> 640 11542.379 13094.544 12251.4450 12085.9211  2504.8292  2213.9972  2186.0784
#> 710 18288.324 19057.604 19274.4961 19150.1301  8159.8105  7840.7013  7982.1704
#> 730 18572.624 18360.710 19410.9788 19600.0305  8881.2575  8580.2444  8730.8894
#> 740 17993.200 17234.786 18345.0370 18576.0659  9585.2310  9315.3147  9476.5872
#> 800 17411.415 17650.343 16899.8006 16708.2290  9543.5866  9184.7365  9262.8004
#>            220        225        230        235        255        300
#> 2    6179.9545  6618.5621  6106.1213  5751.4614  6728.7105  7142.7364
#> 41   7363.8406  7760.6695  6907.0644  6438.8341  8108.3975  8400.4201
#> 42   7190.9803  7581.4827  6705.1383  6231.7589  7950.6767  8228.4775
#> 70   9208.8825  9647.1039  9079.5489  8685.9856  9741.8035 10167.8591
#> 90   9004.7112  9429.3776  8708.4048  8270.1798  9645.5637 10015.5795
#> 91   8808.8518  9227.3880  8466.3814  8017.8740  9478.7142  9829.8627
#> 92   8984.6597  9405.9273  8660.7727  8215.9588  9641.8744 10001.5631
#> 93   8900.1055  9313.5212  8521.5316  8064.5856  9589.1120  9926.9015
#> 100  8632.1912  8995.6903  8024.3394  7529.2295  9434.4664  9664.3588
#> 101  7620.4979  7978.0496  6999.4712  6504.1897  8433.6619  8649.6995
#> 130  9356.3419  9720.1640  8744.6861  8248.3261 10156.1310 10388.7819
#> 135 10246.0883 10566.8548  9503.8744  8999.6999 11091.6168 11250.1692
#> 140  9144.4102  9253.9298  8116.4126  7691.2121  9987.7024  9854.7727
#> 145 10014.6365 10281.4809  9160.3925  8662.5763 10888.8237 10961.9288
#> 150 10139.5293 10319.7715  9167.0904  8701.1001 11013.4453 10965.3908
#> 155 11627.3205 11832.5359 10681.8629 10205.6933 12506.2185 12489.0932
#> 160 11038.9992 11183.8397 10035.3238  9587.9534 11901.2602 11805.1182
#> 165 10931.2348 11062.6862  9917.7985  9478.0104 11787.3126 11674.6813
#> 200   341.1567   751.0537  1263.9856  1585.2040   933.5904  1237.8726
#> 210   428.4049   631.6566  1481.5680  1863.8269   576.9399   936.1867
#> 211   261.7774   492.3778  1315.6183  1711.6983   654.0341   918.0077
#> 220     0.0000   440.7389  1054.2331  1453.9028   879.8307  1037.5667
#> 225   440.7389     0.0000  1152.6942  1629.6079   755.1075   684.4203
#> 230  1054.2331  1152.6942     0.0000   504.1773  1871.3271  1811.9493
#> 235  1453.9028  1629.6079   504.1773     0.0000  2314.9543  2302.1214
#> 255   879.8307   755.1075  1871.3271  2314.9543     0.0000   524.4184
#> 300  1037.5667   684.4203  1811.9493  2302.1214   524.4184     0.0000
#> 325  1107.8841   685.5587  1365.2122  1866.0395  1182.3310   763.8004
#> 345  1450.5113  1034.7918  2033.3608  2537.3009  1000.8402   490.1289
#> 350  2101.1954  1660.4712  2374.1258  2858.6352  1803.2287  1282.3167
#> 360  1876.3081  1474.1334  2477.7399  2981.9115  1296.3793   857.7425
#> 365  2169.6922  2079.2935  3196.3027  3623.2860  1329.0889  1587.0961
#> 380  1547.5534  1549.8985  2598.0794  2993.7144   813.4683  1244.8737
#> 390  1028.5805  1036.8140  2075.3852  2480.4224   355.9538   871.6736
#> 530  5570.7406  5141.1110  5442.6345  5790.5732  5342.5430  4826.4875
#> 600  1814.8866  1870.2791   761.0212   560.6117  2611.6485  2487.9132
#> 630  4222.6840  3830.7328  4788.0483  5287.8404  3514.5232  3186.6161
#> 640  2262.0293  1839.4699  2744.8394  3245.0258  1739.3877  1277.8131
#> 710  8235.8525  8094.7392  9242.0189  9687.3897  7374.7924  7478.1633
#> 730  8989.5971  8887.0472 10020.2482 10442.5490  8149.4602  8297.8249
#> 740  9738.3411  9694.5031 10789.6406 11169.6821  8941.9590  9154.5203
#> 800  9456.6045  9134.0048 10195.6048 10699.7785  8613.1345  8450.8977
#>            325        345        350        360        365        380
#> 2    7234.6844  7608.9740  8273.4664  7997.7781  7214.9679  6654.1071
#> 41   8237.4605  8794.6110  9276.8453  9234.0354  8923.1284  8274.9656
#> 42   8043.6041  8613.6919  9077.9212  9055.6161  8801.5214  8144.9463
#> 70  10256.2461 10636.9574 11299.6388 11020.6417 10100.7895  9599.0333
#> 90   9979.8934 10455.1194 11033.8771 10871.9086 10210.9513  9637.2830
#> 91   9758.4894 10258.4504 10810.5298 10682.7841 10106.4152  9512.7024
#> 92   9945.1658 10435.0117 10998.2351 10856.2185 10239.9187  9655.9121
#> 93   9828.8931 10347.0226 10877.8453 10776.1903 10253.3782  9649.1761
#> 100  9387.7506 10011.2331 10382.7204 10459.4631 10360.6543  9691.8610
#> 101  8363.1535  8990.0862  9358.0131  9438.6197  9403.1227  8725.3227
#> 130 10108.8288 10735.1562 11098.7390 11183.4886 11064.4363 10400.4425
#> 135 10857.4260 11535.7772 11768.1305 11981.5219 12123.9858 11440.2464
#> 140  9179.4494  9891.0993  9706.8646 10246.4245 11310.6455 10682.3908
#> 145 10468.3571 11178.2993 11284.4974 11609.3102 12050.0044 11357.3701
#> 150 10351.6919 11074.7108 10995.1819 11464.9578 12293.1924 11622.0130
#> 155 11893.3697 12616.8515 12547.0981 13012.4524 13752.7191 13068.3773
#> 160 11148.3805 11864.4391 11692.1677 12227.3922 13207.0506 12551.8806
#> 165 11006.5037 11719.2298 11525.1817 12074.3299 13100.9056 12454.2882
#> 200  1433.8758  1694.2508  2394.3745  2095.2433  2105.3779  1437.3864
#> 210  1294.2112  1415.6560  2163.6266  1789.2937  1775.0048  1129.8899
#> 211  1172.8665  1376.6605  2092.0396  1775.3262  1914.7662  1286.1218
#> 220  1107.8841  1450.5113  2101.1954  1876.3081  2169.6922  1547.5534
#> 225   685.5587  1034.7918  1660.4712  1474.1334  2079.2935  1549.8985
#> 230  1365.2122  2033.3608  2374.1258  2477.7399  3196.3027  2598.0794
#> 235  1866.0395  2537.3009  2858.6352  2981.9115  3623.2860  2993.7144
#> 255  1182.3310  1000.8402  1803.2287  1296.3793  1329.0889   813.4683
#> 300   763.8004   490.1289  1282.3167   857.7425  1587.0961  1244.8737
#> 325     0.0000   723.5022  1054.1608  1141.0129  2349.3298  1978.2631
#> 345   723.5022     0.0000   806.5699   448.8939  1806.6445  1623.8049
#> 350  1054.1608   806.5699     0.0000   743.2289  2485.8828  2409.1387
#> 360  1141.0129   448.8939   743.2289     0.0000  1748.3688  1745.1952
#> 365  2349.3298  1806.6445  2485.8828  1748.3688     0.0000   692.7046
#> 380  1978.2631  1623.8049  2409.1387  1745.1952   692.7046     0.0000
#> 390  1532.0925  1330.8018  2137.2702  1576.3249  1152.3520   524.4291
#> 530  4462.9037  4341.7950  3546.1835  4111.6176  5694.8342  5850.0636
#> 600  1903.2735  2621.2438  2777.7571  3043.7665  3940.5477  3355.7349
#> 630  3422.8422  2802.0675  2474.5027  2357.0351  3096.5460  3567.2004
#> 640  1379.6304   812.6897   562.7857   446.9313  2107.0729  2175.4605
#> 710  8143.5948  7430.6634  7631.7422  7076.4243  6067.4573  6722.4305
#> 730  8990.2088  8287.9411  8536.7128  7953.3980  6824.6416  7451.2430
#> 740  9880.6795  9203.0426  9530.2893  8903.5561  7615.3080  8191.6771
#> 800  8841.9389  8163.8305  7932.4189  7717.9913  7592.0128  8273.8103
#>            390       530        600       630        640        710        730
#> 2    6528.7836 11540.680  6159.3962 10202.901  8418.0750 11169.6166 11189.0072
#> 41   8008.0989 11879.512  6653.7999 11586.530  9586.8792 13469.0679 13474.4021
#> 42   7862.1455 11633.807  6432.9818 11411.407  9400.4647 13507.0425 13556.1140
#> 70   9523.5552 14440.423  9032.5682 13166.196 11444.6695 12475.2782 12068.4426
#> 90   9476.8255 13854.899  8545.0124 13154.881 11266.5898 13357.2858 13029.4632
#> 91   9326.2157 13542.878  8272.6753 12993.251 11066.8619 13546.3403 13266.1635
#> 92   9481.8197 13759.068  8477.6864 13155.228 11244.9490 13506.3279 13191.3395
#> 93   9447.2741 13517.897  8299.8963 13101.831 11152.0076 13784.4826 13504.1242
#> 100  9377.1485 12435.441  7635.2098 12806.691 10765.8498 14959.4784 14855.5015
#> 101  8390.3686 11526.722  6614.7530 11783.127  9741.5918 14410.8181 14490.7277
#> 130 10094.1867 13018.375  8343.2331 13529.252 11487.5405 15356.5996 15117.5377
#> 135 11080.0693 13018.386  8994.4447 14242.603 12223.9538 16648.0133 16310.2207
#> 140 10171.6281  9589.747  7383.8605 11858.921 10265.9536 17321.6072 18133.6962
#> 145 10937.4034 12093.935  8567.3098 13721.893 11786.9821 17371.0012 17243.5680
#> 150 11141.3285 11082.688  8479.8737 13261.920 11542.3789 18288.3240 18572.6243
#> 155 12608.2989 12363.843 10010.7684 14782.969 13094.5437 19057.6039 18360.7101
#> 160 12056.3100 11256.424  9320.6401 13795.149 12251.4450 19274.4961 19410.9788
#> 165 11953.0344 11029.782  9195.3570 13596.081 12085.9211 19150.1301 19600.0305
#> 200   957.9557  5892.087  2015.9849  4408.820  2504.8292  8159.8105  8881.2575
#> 210   622.8365  5698.071  2242.5219  4073.370  2213.9972  7840.7013  8580.2444
#> 211   768.8093  5609.190  2075.9597  4092.931  2186.0784  7982.1704  8730.8894
#> 220  1028.5805  5570.741  1814.8866  4222.684  2262.0293  8235.8525  8989.5971
#> 225  1036.8140  5141.111  1870.2791  3830.733  1839.4699  8094.7392  8887.0472
#> 230  2075.3852  5442.634   761.0212  4788.048  2744.8394  9242.0189 10020.2482
#> 235  2480.4224  5790.573   560.6117  5287.840  3245.0258  9687.3897 10442.5490
#> 255   355.9538  5342.543  2611.6485  3514.523  1739.3877  7374.7924  8149.4602
#> 300   871.6736  4826.488  2487.9132  3186.616  1277.8131  7478.1633  8297.8249
#> 325  1532.0925  4462.904  1903.2735  3422.842  1379.6304  8143.5948  8990.2088
#> 345  1330.8018  4341.795  2621.2438  2802.067   812.6897  7430.6634  8287.9411
#> 350  2137.2702  3546.184  2777.7571  2474.503   562.7857  7631.7422  8536.7128
#> 360  1576.3249  4111.618  3043.7665  2357.035   446.9313  7076.4243  7953.3980
#> 365  1152.3520  5694.834  3940.5477  3096.546  2107.0729  6067.4573  6824.6416
#> 380   524.4291  5850.064  3355.7349  3567.200  2175.4605  6722.4305  7451.2430
#> 390     0.0000  5663.392  2832.0115  3679.710  2023.2557  7218.0211  7962.1235
#> 530  5663.3923     0.000  5412.6689  3219.736  3674.9055  8318.3015  9251.7310
#> 600  2832.0115  5412.669     0.0000  5250.518  3235.2083  9960.0532 10757.2042
#> 630  3679.7101  3219.736  5250.5181     0.000  2043.2118  5613.3057  6568.8613
#> 640  2023.2557  3674.905  3235.2083  2043.212     0.0000  7070.2362  7973.9849
#> 710  7218.0211  8318.302  9960.0532  5613.306  7070.2362     0.0000   958.3579
#> 730  7962.1235  9251.731 10757.2042  6568.861  7973.9849   958.3579     0.0000
#> 740  8714.3376 10410.300 11545.4784  7682.338  8969.3308  2104.0755  1158.5792
#> 800  8628.5698  6742.641 10705.5341  5458.068  7475.5769  3289.9191  3719.5314
#>           740       800
#> 2   10926.086 14170.901
#> 41  13121.384 16348.957
#> 42  13250.033 16291.254
#> 70  11319.145 15756.301
#> 90  12336.532 16642.803
#> 91  12614.573 16810.403
#> 92  12506.962 16787.236
#> 93  12845.603 17045.326
#> 100 14327.538 17891.319
#> 101 14180.903 16991.178
#> 130 14443.853 18503.454
#> 135 15493.886 19700.806
#> 140 18560.242 16075.995
#> 145 16534.338 18777.986
#> 150 17993.200 17411.415
#> 155 17234.786 17650.343
#> 160 18345.037 16899.801
#> 165 18576.066 16708.229
#> 200  9585.231  9543.587
#> 210  9315.315  9184.736
#> 211  9476.587  9262.800
#> 220  9738.341  9456.604
#> 225  9694.503  9134.005
#> 230 10789.641 10195.605
#> 235 11169.682 10699.779
#> 255  8941.959  8613.134
#> 300  9154.520  8450.898
#> 325  9880.679  8841.939
#> 345  9203.043  8163.830
#> 350  9530.289  7932.419
#> 360  8903.556  7717.991
#> 365  7615.308  7592.013
#> 380  8191.677  8273.810
#> 390  8714.338  8628.570
#> 530 10410.300  6742.641
#> 600 11545.478 10705.534
#> 630  7682.338  5458.068
#> 640  8969.331  7475.577
#> 710  2104.075  3289.919
#> 730  1158.579  3719.531
#> 740     0.000  4607.829
#> 800  4607.829     0.000

# }